refine test
